# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD de Produtos - Documentação da API",
        "description": "Documentação da API para o sistema CRUD de produtos",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os produtos",
                "description": "Retorna uma lista paginada de todos os produtos",
                "operationId": "getProdutosList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Produto"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar um novo produto",
                "description": "Armazena um novo produto no banco de dados",
                "operationId": "storeProduto",
                "requestBody": {
                    "description": "Dados do produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "nome": [
                                                    "O campo nome é obrigatório."
                                                ],
                                                "preco": [
                                                    "O campo preço é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Obter um produto específico",
                "description": "Retorna um produto específico pelo ID",
                "operationId": "getProdutoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar um produto existente",
                "description": "Atualiza um produto existente no banco de dados",
                "operationId": "updateProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do produto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProdutoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "nome": [
                                                    "O campo nome é obrigatório."
                                                ],
                                                "preco": [
                                                    "O campo preço deve ser numérico."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Excluir um produto",
                "description": "Exclui um produto existente",
                "operationId": "deleteProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto excluído com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProdutoRequest": {
                "title": "ProdutoRequest",
                "description": "Dados para criar ou atualizar um produto",
                "required": [
                    "nome",
                    "preco"
                ],
                "properties": {
                    "nome": {
                        "description": "Nome do produto",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Smartphone XYZ"
                    },
                    "descricao": {
                        "description": "Descrição do produto",
                        "type": "string",
                        "example": "Smartphone com 128GB de armazenamento e 8GB de RAM",
                        "nullable": true
                    },
                    "preco": {
                        "description": "Preço do produto",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1299.99
                    },
                    "categoria": {
                        "description": "Categoria do produto",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Eletrônicos",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Produto": {
                "title": "Produto",
                "description": "Modelo de produto",
                "properties": {
                    "id": {
                        "description": "ID único do produto",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nome": {
                        "description": "Nome do produto",
                        "type": "string",
                        "example": "Smartphone XYZ"
                    },
                    "descricao": {
                        "description": "Descrição do produto",
                        "type": "string",
                        "example": "Smartphone com 128GB de armazenamento e 8GB de RAM"
                    },
                    "preco": {
                        "description": "Preço do produto",
                        "type": "number",
                        "format": "float",
                        "example": 1299.99
                    },
                    "categoria": {
                        "description": "Categoria do produto",
                        "type": "string",
                        "example": "Eletrônicos"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-11T18:30:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-11T18:30:00.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Produtos",
            "description": "Endpoints da API para gerenciamento de produtos"
        }
    ]
}